04_ArtificialSynthetic_datasets_generation_and_the_second_deep_learning_training/requirements.txt
numpy==1.21.3
torch==1.10.0
torchvision==0.11.1
opencv
labelme
Pillow
